Divide.simplify(Arguments)
Now simplifies division expressions when second argument is minus one.
e.g. (/ 4 -1) now simplifies to -4.

+++ src/main/java/org/oakgp/function/math/LongUtils.java
@@ -131,7 +131,7 @@ public final class LongUtils implements NumberUtils {
       return new ConstantNode(i, TYPE);
    }
 
-   private int evaluate(Node n) {
+   private long evaluate(Node n) {
       return n.evaluate(null);
    }
 }
